+++ C++/Algorithms/hashing.cpp
