+++ C++/Basic Problems in C++/34_Messages_Order.cpp
