+++ C++/Basic Problems in C++/7.cpp
@@ -1,0 +1,8 @@
+bool search(vector<int>& nums, int target) {
+        for(int i=0;i<nums.size();i++)
+        {
+            if(target==nums[i])
+                return true;
+        }
+        return false;
+    }
