+++ C++/Basic Problems in C++/ClockReset
