+++ C++/Basic Problems in C++/Count_of_Subarrays.cpp
