+++ C++/Basic Problems in C++/Get_Minimum_Squares.cpp
