+++ C++/Basic Problems in C++/LCM.cpp
