+++ C++/Basic Problems in C++/LinearSearch.cpp
