+++ C++/Basic Problems in C++/Maximize_Toys.cpp
