+++ C++/Basic Problems in C++/Merge two sorted arrays
