+++ C++/Basic Problems in C++/NumberIntoEnglishWordClass.cpp
