+++ C++/Basic Problems in C++/Resetting Clock
