+++ C++/Basic Problems in C++/Reverse number
