+++ C++/Basic Problems in C++/Reverse_integer.cpp
