+++ C++/Basic Problems in C++/Reverse_integer.exe
