+++ C++/Basic Problems in C++/databaseManagement.cpp
