+++ C++/Basic Problems in C++/maxDifference.cpp
