+++ C++/Basic Problems in C++/pattern.cpp
