+++ C++/Basic Problems in C++/stack.cpp
