+++ C++/Basic Problems in C++/stockBuyAndSell.cpp
