+++ C++/Basic Problems in C++/stockBuyAndSell.exe
